Initial cut of FreeBSD support.
This is a first step for making devstack work on FreeBSD 11 host.  Current
configuration only uses nova-compute component.

There's a lot of loose ends still and areas to be fixed and developed.

Conflicts:
	functions
	lib/cinder
	lib/swift
	stack.sh
	stackrc
	tools/create-stack-user.sh

+++ stack.sh
@@ -142,7 +142,7 @@ disable_negated_services
 
 # Warn users who aren't on an explicitly supported distro, but allow them to
 # override check and attempt installation with ``FORCE=yes ./stack``
-if [[ ! ${DISTRO} =~ (precise|saucy|trusty|7.0|wheezy|sid|testing|jessie|f19|f20|rhel6) ]]; then
+if [[ ! ${DISTRO} =~ (precise|saucy|trusty|7.0|wheezy|sid|testing|jessie|f19|f20|freebsd11|rhel6) ]]; then
     echo "WARNING: this script has not been tested on $DISTRO"
     if [[ "$FORCE" != "yes" ]]; then
         die $LINENO "If you wish to run this script anyway run with FORCE=yes"
@@ -182,23 +182,24 @@ if [[ $EUID -eq 0 ]]; then
     exit 1
 fi
 
-# We're not **root**, make sure ``sudo`` is available
-is_package_installed sudo || install_package sudo
+if [[ $(whoami) != $STACK_USER ]]; then
+    err "Current user: $(whoami).  Expected: $STACK_USER. Please do 'su - $STACK_USER' and rerun."
+fi
 
-# UEC images ``/etc/sudoers`` does not have a ``#includedir``, add one
-sudo grep -q "^#includedir.*/etc/sudoers.d" /etc/sudoers ||
-    echo "#includedir /etc/sudoers.d" | sudo tee -a /etc/sudoers
+# XXX this is inherently broken: if we are not root and not in sudoers already
+# modifying /etc/... will fail due to no access.  Installing sudo by non
+# privileged user will also fail etc.
 
 # Set up devstack sudoers
-TEMPFILE=`mktemp`
-echo "$STACK_USER ALL=(root) NOPASSWD:ALL" >$TEMPFILE
+#TEMPFILE=`mktemp`
+#echo "$STACK_USER ALL=(root) NOPASSWD:ALL" >$TEMPFILE
 # Some binaries might be under /sbin or /usr/sbin, so make sure sudo will
 # see them by forcing PATH
-echo "Defaults:$STACK_USER secure_path=/sbin:/usr/sbin:/usr/bin:/bin:/usr/local/sbin:/usr/local/bin" >> $TEMPFILE
-echo "Defaults:$STACK_USER !requiretty" >> $TEMPFILE
-chmod 0440 $TEMPFILE
-sudo chown root:root $TEMPFILE
-sudo mv $TEMPFILE /etc/sudoers.d/50_stack_sh
+#echo "Defaults:$STACK_USER secure_path=/sbin:/usr/sbin:/usr/bin:/bin:/usr/local/sbin:/usr/local/bin" >> $TEMPFILE
+#echo "Defaults:$STACK_USER !requiretty" >> $TEMPFILE
+#chmod 0440 $TEMPFILE
+#sudo chown root:root $TEMPFILE
+#sudo mv $TEMPFILE /etc/sudoers.d/50_stack_sh
 
 # Additional repos
 # ----------------
@@ -248,6 +249,7 @@ fi
 # Create the destination directory and ensure it is writable by the user
 # and read/executable by everybody for daemons (e.g. apache run for horizon)
 sudo mkdir -p $DEST
+# XXX This fails if TRACK_DEPENDS=True
 safe_chown -R $STACK_USER $DEST
 safe_chmod 0755 $DEST
 
@@ -625,7 +627,7 @@ set -o errexit
 
 # Print the commands being run so that we can see the command that triggers
 # an error.  It is also useful for following along as the install occurs.
-set -o xtrace
+#set -o xtrace
 
 
 # Install Packages
@@ -638,9 +640,12 @@ set -o xtrace
 echo_summary "Installing package prerequisites"
 source $TOP_DIR/tools/install_prereqs.sh
 
-# Configure an appropriate python environment
-if [[ "$OFFLINE" != "True" ]]; then
-    $TOP_DIR/tools/install_pip.sh
+# XXX why not install pip from ports/pkg?
+if ! is_freebsd; then
+    # Configure an appropriate python environment
+    if [[ "$OFFLINE" != "True" ]]; then
+        $TOP_DIR/tools/install_pip.sh
+    fi
 fi
 
 # Do the ugly hacks for borken packages and distros
@@ -994,13 +999,18 @@ fi
 # ----
 
 if is_service_enabled n-net q-dhcp; then
-    # Delete traces of nova networks from prior runs
-    # Do not kill any dnsmasq instance spawned by NetworkManager
-    netman_pid=$(pidof NetworkManager || true)
-    if [ -z "$netman_pid" ]; then
-        sudo killall dnsmasq || true
-    else
-        sudo ps h -o pid,ppid -C dnsmasq | grep -v $netman_pid | awk '{print $1}' | sudo xargs kill || true
+    # TODO we have to reinit dhcpd here for FBSD
+    if ! is_freebsd; then
+        # Delete traces of nova networks from prior runs
+        # Do not kill any dnsmasq instance spawned by NetworkManager
+        netman_pid=$(pidof NetworkManager || true)
+        if [ -z "$netman_pid" ]; then
+            sudo killall dnsmasq || true
+        else
+            sudo ps h -o pid,ppid -C dnsmasq | grep -v $netman_pid | awk '{print $1}' | sudo xargs kill || true
+        fi
+
+        clean_iptables
     fi
 
     clean_iptables
@@ -1012,7 +1022,11 @@ if is_service_enabled n-net q-dhcp; then
     fi
 
     # Force IP forwarding on, just in case
-    sudo sysctl -w net.ipv4.ip_forward=1
+    if is_freebsd; then
+        sudo sysctl net.inet.ip.forwarding=1
+    else
+        sudo sysctl -w net.ipv4.ip_forward=1
+    fi
 fi
